Initialize README.md, 0000-template.md from Rust RFCs
Sawtooth is adopting a process which is very similar to the process
in place for Rust RFCs.  The Rust RFCs repository is at:

  https://github.com/rust-lang/rfcs

The README.md and 0000-template.md are copied from the above repository
and copyright of the text can be determined in that repository.  The
repository is dual-licensed, and we are using the README and template
under Apache 2.0.

The LICENSE-APACHE file from the above repository is included.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>