Add RFC for Command Family
The Command Family is a transaction family that can exercise all of the
operations provided on the TransactionContext in addition to other
actions, such as sleeping or marking the transaction as invalid.  This
family will be used for creating a wide variety of tests, both for
correctness and performance of the Transact library.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>